MB-64349 Test Encryption Key Settings Button
Adds a new button in the add/edit encryption key dialog:
"Test Encryption Key Settings"
The button calls a validation endpoint with the
form values.
The result is displayed to the right of the button
in the specific success/error colour.
For each error there will be an error text under the specific form field.

Change-Id: I2fdbb1abfa585c6c3b8bb44031decb09e052c559
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222516
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>