MB-59704: read cgroup_base from profile
I incorrectly used the fallback default path as the main path instead
of using the profile's entry. This was found because setting it up with
docker requires some extra configuration that involves overriding the
cgroup_base_path. When it failed to walk the folders because of
permissions, it was obvious it was walking from the incorrect location.

Change-Id: I041e9f65392d36e4b8ec53239d7334c635d6486a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222800
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>