MB-64587: Switch to sans-io and test interleaved pings
We need to support receiving (and handling) pings in the middle of
a fragmented message. In order to test that functionality, we can't use
the simple sync websocket client. By using sans-io (ClientProtocol), we
can add a ping frame in the middle of a fragmented message.

Change-Id: Ie4ce1119fc9164616ffe3529d69adab283d29451
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222840
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>