MB-64812: [FBR] Add file based backfill support
Add support for a basic file based backfill. This is not the complete
implementation, nor will it work for all scenarios.

File based backfill will:

1) Issue a DownloadSnapshot command for some vBucket to the destination
   memcached.
2) Poll stats on the destination to determine when the DownloadSnapshot
   command completes.
3) Issue a SetVBucketState message instructing memcached to warm up the
   vBucket from the previously transferred snapshot.

Snapshots are released by the janitor, which was implenented in a
previous change. Use of the janitor for this means that they will only
be removed outside of rebalance (or at the start of one). A DCP
stream will be set up at step 3 for the vBucket in question.

This patch does not:

a) Deal with encyption at rest.
b) Run DCP and File based backfill in parallel.

Change-Id: Ib28a9c1dbcb1fa7cb1fb44303d01adbf88521f48
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222847
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>