MB-64812: [FBR] Release snapshots in janitor
We (ns_server) "own" the snapshots that we instruct KV to create. By
that, we mean, that we are responsible for their destruction too. This
comes about as KV would otherwise need to release snapshots after some
period of inactivity, and ns_server ultimately has knowledge of when
snapshots are no longer useful.

To handle rebalance failures, graceful or not, ns_server deletes
snapshots via the janitor. Any snapshot found by the janitor will be
removed.

Change-Id: I50fa7d396678177bcf80945fbe2ae3713d9252eb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223388
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>