MB-61292: Remove NULL_DEK from deks_being_dropped...
... if get_ids_in_use_callback returns a list that
doesn't contain NULL_DEK (which basically means NULL_DEK
is not used anymore).
For other deks, we keep them in deks_being_dropped
until we remove the dek, but in case of NULL_DEK there is
nothing to remove so it should be safe to remove it from
deks_being_dropped immediately when we know that it is not
used. Otherwise it will get to list of keys to be dropped
next time again.

Scenario that demonstrates the issue:
1. Enable encryption for a bucket
2. Configure short lifetime for DEKs with long DEKs rotation interval
3. Wait for DEK to become expired, (and maybe restart the node, not
   sure if it is required)
4. Node will try to drop NULL_DEK because HasUnencryptedData is
   undefined, and after that it will run compations trying to
   drop NULL_DEK in a circle (because NULL_DEK never leaves the
   deks_being_dropped list).

Change-Id: I2de3f931af29ae24f309e276353ea31345a8b4fe
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222936
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>