MB-65110 MB-61292: Add active key to the deks_being_dropped
If active key expires, we should not drop it, but we should
add it to the list of keys being dropped.
This is needed so that we can avoid retrying immediately.
Basically if we don't do that, the cleanup timer fires
every second, which will polute the logs (debug and event log).
If we do that, we retry in DEK_DROP_RETRY_TIME_S seconds
(which is 3 hours by default).

Change-Id: I9cccd0c402d54e86cc721dc217266e5317eb2882
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222939
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>