MB-64587: Start ns_server_stats before menelaus_sup
Otherwise when we add cm_app_telemetry_curr_connections and try to
initialise it to 0 on start, ns_server_stats wouldn't yet be up, so
the initial value wouldn't be recorded.

This means now that certain report stat calls could initially fail, as
ns_server_stats is now started before ns_audit and menelaus_cbauth.
However, this is fine as these calls are guarded by `try`s, so no other
stats would be affected, and the time window for this is very small.

Other processes that are now started after ns_server_stats are not
expected to cause any issues, as they were not found to be used in
ns_server_stats.

Change-Id: Ic9b04ab995b4f6d6be9d0edbd8e5a390de5693fa
Reviewed-on: https://review.couchbase.org/c/ns_server/+/222990
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>