MB-64587: Reduce AppTelemetryTests.simple_test run time
The typical runtime previously was about 5s, but this was because we
waited for prometheus to scrape the metrics, which can add a couple
seconds, and we also only perform the app telemetry scrape once a
second, which can add another couple seconds.

By directly fetching the metrics from /metrics, and setting the
app telemetry scrape interval to 0, we get the run time consistently
less than a second (typically 0.13s locally), which is much more
reasonable.

Change-Id: I42a274ac966341959e75ea0f0fc8412d861a1890
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223082
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>