MB-63853: enable useRelativeWebRedirects by ...
... default in morpheus. The easiest way to do this is to simply
change what the default value is based on the cluster_compat
is_cluster_morpheus call. Alternatively you could remove the un-guarded
key and replace it with a new key guarded by the cluster-compat mode but
this seemed undesirable for a few reasons. The main one being that we
would likely have to rename the key. That's why this method seemed
better.

Change-Id: I280f69f337a91c9fb4f3084054fd0c0c57c4bb58
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223042
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>