MB-64587: Fix disconnect_on_disable_test
While the test currently correctly passes, if the node doesn't
disconnect the client when app telemetry is disabled, the test will
still pass because it would receive a close message after the first
scrape times out.

The test should first respond to the first scrape, then expect the
close message. If the server doesn't disconnect the client, we will
catch this by either getting a non-close frame, or a timeout waiting
for the next frame.

Because this test follows the websocket closing handshake, by
automatically responding with CLOSE after receiving a CLOSE (in
WebsocketConnection._get_and_send_data), it is possible to get a
BrokenPipeError, because mochiweb_websocket currently drops the
connection as soon as it sends the CLOSE, rather than waiting to
receive a CLOSE. This is handled with a try/except for now, but this
will be removed once closing handshake is correctly implemented.

Change-Id: If3650e6fa8ce42986d8eeb4355982d495f292c6d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223094
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>