MB-63210 - remove placeholder text in bucket details summary
The initial requirement for MB-63210, making Magma the default storage
engine, suggested two placeholder names for 128 and 1024
vBuckets. Eventually it was decided to refer to the two configurations
by the vBucket count. However this change was not reflected in bucket
detailed view, which still referred to "Configuration 1" and
"Configuration 2". This change removes that obsolete text. The
detailed view already showed the number of vBuckets in use.

Change-Id: I95b3c730b6426aec23072301c49767cd3442617b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223152
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>