MB-65260: Fix cgroup validator nonnegative integer
There was a small hole in the validators which did not ensure the value
is equal to or above 0. This fix changes that and requires that the
integer is greater than or equal to zero and provides a nice error
message if it's outside that range.

Change-Id: I31d59e4b2a3af071b33bad1f3ce75c9eddd54065
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223228
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>