MB-64812: [FBR] Refactor single vb mover for new backfill type
Move the DCP backfill code to a new function such that we can choose
whether or not to run it when we introduce a new type of backfill.

Two small other changes have been made; the log message on backfill
completion has been updated to be less specific, file based backfill
will not wait for a high sequence number to be persisted, and the
backfill phase start event has been moved earlier as we won't set the
vBucket state in the same way for file based backfills.

Change-Id: I14d745712a0f185d6d4d24df3a1cb4a8caa95109
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223389
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>