MB-64910 MB-61292: Make sure old DEKs are removed...
... when a new bucket is created

Scenario:
1. Create 2 node cluster with a bucket
2. Remove node#2
3. Readd node#2 to the cluster.

At step #3 node#2 will still have the bucket directory with old DEKs
(from the time when it was part of the cluster). It will cause errors
with update of integrrity keys (because that old DEK will not be
readable).
Since this DEKs is useless anyway, it seems like it makes sense to
just garbage collect it.

Change-Id: I89eb5f21783a54f4ef6cd625ad520d740cd4e44e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223327
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>