MB-64910 MB-61292: Don't update integrity proof for deks...
... that no longer exist

It can happen that node contains the bucket directory
but this node is not KV node anymore. Node can still try
to update the integrity proof keys in this case, but it will
fail because such DEKs are not maintained anymore (so it is
not guaranteed that we will be able to decrypt them).

Change-Id: I9dba64a0eeaf4bb7f04ac72146d5af0af2e81db5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223329
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>