MB-65178: Allow startup even if unused keks can't be written locally
With this change, if a key is unused, we try to write it locally
to the node disk but don't prevent startup if it can't be written.
This allows us to startup even if there are failures to write keys
to disk that are not used by anything. For example, if an unused
KMIP key can't be written locally because of no connection to server,
we still start up. The periodic job to ensure that all keks
are on disk will keep on retrying to write the unwritten ones to disk
after server start.

Change-Id: I15afed8472f5129e91d14ce8fd58073ebfa35354
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223333
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>