MB-61292: [cluster tests] Sort nodes when verifying services
Nodes can change order in the connected_nodes list (e.g. if we remove
and add node back). In this case the services requirement can fail
because it depends on the order of nodes in the list.
This change makes the requirement agnostic to that order.

Change-Id: I4f7db19fc51078937d82066fc08e94c4dd177eb9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223445
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>