MB-65125 Security Other Settings Conversion
Convert mn.security.other.components.js from
Angular to React

## Overlay Trigger

I discovered that certain tooltips in OnPrem
require hovering over their contents (e.g. so we
can click on a link in the tooltip).

In react-bootstrap, there is no out-of-the-box
prop to do that - it requires manually setting
state and mouseEnter + mouseLeave. In my next PR
for Security conversion, I will introduce a
wrapper component that has the same interface as
the react-bootstrap OverlayTrigger, but takes in
additional allowContentHover prop which wraps the
OverlayTrigger in an appropriate div and manages
the state.

Change-Id: I4c4a989e63dac20deda86a67d73cf870f03ca9f2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223606
Tested-by: Piotr Najda  <piotr.najda@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>