MB-64587: Use a larger continuation frame to cover error case
When the continuation frame uses larger payload lengths, we hit a bug
where the continuation frame would be considered an incomplete frame,
causing it to get stuck waiting for the supposed rest of the frame.
By using a larger continuation frame, we cover this case.

Change-Id: Idb445a4b29eceb761808f0ee2606c8f9d176b3cb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223621
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>