MB-64915: Fix to refresh cached dek snapshot during log collection
During encrypted log collection, we take the dek snapshot, cache it
and use it to decrypt all rotated files and the most recent file of a
file type. It is possible that dek got rotated after we had cached the
dek snapshot, and the log file got rotated as result, and the dek got
garbage collected because it was no longer in use anymore. In such a
case the old DS won't be able to decrypt the file.

This change will update the cached DS to the latest if it finds the
current file can't be decrypted with the current DS.

Change-Id: I92bb1e1b8abda2d89160994bf62af08dadd49a36
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223789
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>