MB-65461: Aggregate user activity when fetching user details
Rather than having to wait for the activity aggregation interval,
before activity is included in the user details, the activity should
be aggregated whenever the user details are returned.

This also ensures that we don't need to sync_activity before each
check in the associated UsersTestSet tests.

Change-Id: I63f66ccccab332c2df55fddc80722bb21bf8fb40
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223893
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>