MB-65461: Don't update aggregated activity when unchanged
If we update aggregated activity for each user even if it hasn't
changed, this could cause retrieving user details to be quite costly,
now that we aggregate the activity when retrieving user details.

If we first check if the timestamp has changed, we can avoid most of
the replicated_dets updates, that would otherwise be caused when the
user details are being regularly fetched.

Change-Id: I9cbfaeed7417e533fe5ef20c104ffffc9cb0b742
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223894
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>