MB-65464: Add settings for stuck DCP pruning
Add two settings to configure DCP connection pruning. A corresponding
kv_engine change will be made to support these settings, however, it
is not necessary to merge at the same time as ns_server will handle
lack of support from kv_engine gracefully.

A regex can be specified which will enable this feature, and tell
memcached which connections (matching based on their name) ought to be
considered for pruning. This setting is called
`dcp_disconnect_when_stuck_name_regex`. This setting is passed to
memcached as a base64 encoded string. This allows us to avoid escaping
concerns with respect to quotation marks when we pass the string in
memcached.json. To handle older versions nodes existing in the cluster,
this configuration is stored in it's base64 encoding by the REST API,
which will decode it before returning any value to the user.

A timeout value (seconds of "stuckness" after which a disconnect will
happen) can be specified as a positive integer. This setting is called
`dcp_disconnect_when_stuck_timeout_seconds`.

This commit is a cherry-pick of the original 7.2.3 commit as this is
being pushed to a maintenance release. We cannot merge forwards to this
branch as changes updated in the neo branch are not necessarily present
in this version of 7.6. The type ?MC_MAX_INT was also updated in this
version.

Change-Id: I02faac3bec2e4962afbde94ce1623f371ee90aca
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224142
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>