MB-65336 MB-61292: Increment encr counter in chronicle...
... once per encr method when reencrypting DEKs

For example, if we have 100 DEKs, and all of them needs to be
reencrypted with secret A, we can increment the counter for
secret A only once. Currently we do it 100 times (for each DEK).
This leads to a lot of unnecessary writes to chronicle, which
leads to retries_exceeded errors.

Change-Id: Ica85941bbe564b8e2d8b24edcabb0101ab5541a9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223849
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>