MB-65322: Add option to include user in ui cookie
This functionality is disabled by default.
To enable it, an admin would need to make the following internal API
request:
POST localhost:9000/internalSettings
-d includeUsernameInUICookie=true

As this is an internal API, and is intended to be added in a
maintenance release, there is no cluster compat version check.

If this functionality is enabled, the auth token will be returned with
the base64 encoded username appended, separated by a "-". If ns_server
authenticates a user with a token which has a username appended, it
will confirm that the user matches the token. If no username is
appended, the request will be rejected.

Change-Id: I7951ec3c8ace6e5329dfc6c7f9313faa8978fd92
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224356
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>