MB-65453: guard API's by compat-version properly
During some of the changes in review, the guards were moved around and
I failed to properly protect some of the API endpoints from
cluster-compatibility changes/differences. Part of the issues is some of
the functions such as cgroups:supported/0 get called very early in
initialization so it cannot also contain calls to ns_config. This
requires some splitting of the calls and in that process we failed to
check the cluster-compat on the API endpoints. This patch resolves that
by adding the calls to is_cluster_morpheus and is_enterprise s/t we
don't show the API's anywhere we shouldn't.

The internal setting enableCgroups would still be accessible in
mixed-mode to make pre-configuration easier.

Change-Id: I66558a47447adbc44d4829c68d5d52ed6dd60583
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223921
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>