MB-65336 MB-61292: Send repair_join in ns_cluster:init...
... if chronicle state is removed. Otherwise we are risking to get
stuck in the loop forever.

This can happen if ns_server start fails during join process.
The other node (that sends /completeJoin) will remove this
node from cluster as soon as it receives the completeJoin response
(error). It is possible that if we don't repair join here ns_cluster
will try starting ns_server again and again until it succeeds (only
after that it will leave the cluster). The problem is that it may be
unable to start ns_server (because chronicle is dead already), so it
will be stuck in the loop forever.

Change-Id: Ia1210c957b1f1d439c1bd7d267bae97980ccd6ce
Reviewed-on: https://review.couchbase.org/c/ns_server/+/223946
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>