MB-65373: [goport] Fallback to regular 'kill' ...
... when the kill directed at the process group fails. This can happen
for a few reasons, most of which are benign timing issues. The other
issue we see is when using cgroups, as has been reported. The safest
solution is to just fallback to the regular kill command in these cases
so that we cannot leave any processes orphaned or still running.

Change-Id: Ib1b0bfc110c431bd9af1c4efdf5492077b88c7d3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224153
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>