MB-65528: Add mirror_ssl_port in provisioned_profile
This port will be used for kv bootstrap, when a load balancer is unable
to use the same port for bootstrap and subsequent kv ops.

This will not be exposed in any REST API, or cluster config (such as
that seen in the Cluster Config Carrier Protocol).

It is enabled via the config profile as it is only appropriate for
provisioned clusters, and because this allows us to add it without a
cluster compat mode change, since the ns_config/chronicle config is
unaffected.

Note, support for remapping of this port in cluster_run has not been
added. This would require additional work and testing to validate, and
does not appear necessary. The current behaviour is that the same 11300
port would be used for the mirror_ssl_port in any cluster_run spawned
node where --provisioned has been specified. Because memcached
listens with SO_REUSEADDR, as long as only one cluster is active at
the same address, bootstrapping behaves correctly using this port
(because any node can handle the bootstrapping request, and subsequent
kv ops will just go to the correct non-bootstrap ports).
If more advanced cluster_run testing is required for this feature in
future, then we can add explicit handling for a custom mirror_ssl_port
for cluster_run nodes at that point.

Change-Id: Ic5d4944123f57aacd225278941084bde49727b6d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224178
Tested-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker