MB-65601: Fix race in aggregate_activity_test__
It is possible for the activity to be inserted before the activity
cleanup gets triggered, causing replicated_dets:get(user_storage, ...)
not to match the mecked function. Since this test is only intended to
handle the aggregation side, we don't need to cover this path (which
is covered by tests in activity_tracker).

If we update the configuration before starting activity_tracker, then
the cleanup doesn't occur, so the race can't occcur.

Change-Id: I95ff2c01f403307450073932b3f96ad533e2db87
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224348
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>