MB-65270: fix FFDC collection to properly end file
The FFDC collection task wasn't releasing the final newlines at the end
of the files it processes. This causes issues when they are read later.
After inspecting the exact results of the files it was clear we were
simply eating the final newline. This change fixes it and generally file
tasks like this should end with stream.clear_wraparound().

Change-Id: Ic9b764b578ae24c68501d12f83f36e6b35a6f4d0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224366
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>