MB-65270: [BP] fix FFDC collection
The FFDC collection task wasn't releasing the final newlines at the end
of the files it processes. This causes issues when they are read later.
After inspecting the exact results of the files it was clear we were
simply eating the final newline. This change fixes it and generally file
tasks like this should end with stream.clear_wraparound().

Change-Id: I424b8d9f90f9fdf4786472c3e44b8da099ef9ed8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224792
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker