MB-65620: Configurable app telemetry min/max scrape interval
For testing purposes, it is useful to be able to set the scrape
interval below 60s. For the SDK team's automated tests in particular,
they currently have to have a >60s sleep in the test just to cover
app telemetry, which is obviously wasteful. Their tests do not have the
ability to make requests from localhost on the node, so /diag/eval is
not an option.

This change adds the following parameters to the /internalSettings
endpoint:
- appTelemetryMinScrapeIntervalSeconds
- appTelemetryMaxScrapeIntervalSeconds

Both of these can be set to any positive integer, and there is
no validation that min <= max.

Change-Id: Iedbf838776797960adfb0c24e4aabbab369293a1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224422
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>