Dump config in AFO stale config test to logs
This test fails occasionally, and it appears that the failover is
proceeding as the partitioned off nodes are still in the config. The
config pull should update the config to remove them. Dump the config
after it is updated, to see if we can work out what's going on in the
test when it fails.

Change-Id: I8d0ad0189bda6f428400716d4288a8e96bf99a65
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224428
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>