MB-61292: Allow DEK reencryption to fail in init
Originially the logic was the following:
DEK reencryption in cb_cluster_secrets:init/1 must succeed, because
if it fails after node addition (or cluster leave), old KEKs will be
garbage collected, and after that DEKs will become not decryptable.
KEKs will be garbage collected in this case because during
node addition or cluster leave chronicle config gets replaced
while cb_cluster_secrets process is down. Old KEKs configuration
simply disappears.

For example, node join scenario on node-to-be-added:
1. cb_cluster_secrets shuts down (as part of completJoin/1)
2. chronicle config gets replaced with cluster's config
3. cb_cluster_secrets starts and finds out that all old KEKs
   have disappeared from the config, so it needs to garbage collect
   them

However, this logic doesn't work well for another scenario.
When a node starts, it tries to reencrypt DEKs in
cb_cluster_secrets:init/1, but if that node doesn't have a quorum
it will fail. Basically, that node will never be able to start
until it gets a quorum, which is not good.

New logic:
DEK reencryption in cb_cluster_secrets:init/1 can fail now,
we will retry reencryption later. This resolves the issue with
node start when there is no quorum.

However, for the node join scenario, we still need to make sure
that DEKs are reencrypted successfully before old KEKs are garbage
collected (we can't retry in this case). For this case we explicitly
call cb_cluster_secrets:reencrypt_deks/0 during node join and
cluster leave procedures while the cb_cluster_secrets process is down.
In this case, we can't retry, so check the reencryption returns ok.

Change-Id: I2c39c27453193e05096e29716fc3797d4030ce4a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224435
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>