MB-64871 Encryption At Rest Fixes
Times displayed in Encryption at Rest pages are in local time
with the GMT clearly specified when setting the rotation time
of an encryption key.
Added limits for hour and minutes fields.
24 hour format clarification when setting the rotation time
of an encryption key.
Smaller fields for hour and minute when setting
the rotation time of an encryption key.

Change-Id: I7fafa94389f3dbbd47f248c98bc8d32dada88de8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224489
Tested-by: Raluca Lupu <raluca.lupu@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>