MB-65367 LDAP minor bug fixes
## 1. Advanced Settings cut-off

After moving LDAP from a page to a dialog, we need
to add a bottom margin that accounts for the
absolutely-positioned footer.

Else, when the LDAP section exceeds page height,
(which is the case with Advanced Settings being
at the bottom with other sections above it enabled
) the footer will cause the section to be cut off.

## 2. Clear Cache button

The button was not working because we were using
a dialog dismiss prop, which is not available
on a page.

## 3. Dead RFC link

The link to the RFC was dead. We don't hyperlink
to a specific section now since there's no
way to do that, but the whole RFC is about the
subject matter so it's fine.

Change-Id: I3dd2fca20749e75c2a129061aae4e9c0d6bfd302
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224490
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Piotr Najda  <piotr.najda@couchbase.com>