MB-63720 support fusion uploaders on the ns_bucket level
fusion uploaders are stored at the chronicle key
{bucket, <bucketname>, fusion_uploaders}

fusion uploaders are stored as the list of tuples
{Node, Term} where each tuple corresponds to vbucket
Node is the node that is uploading the vbucket
Term is an integer that is increased every time
when Node changes. The purpose of Term is to help
s3 to recognize rogue uploaders and ignore them.

Change-Id: I5087d265561fdb59e0f2f3c5267118de4e02bd64
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224653
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>