MB-63255: Decrease cores_per_bucket guardrail from 0.4 to 0.2
The default value was originally set to 0.4, as this was the value
which there was sufficient confidence in. There is now confidence that
0.2 cpu cores for each bucket is enough, so we can update the default.

This doesn't impact self-managed clusters, as this guardrail is not
configured for self-managed.

When a provisioned cluster is upgraded from trinity, the minimum will
be changed only if it is still set to the old default of 0.4, otherwise
it will not be changed.

Change-Id: I83fe5ad02571d7b8fe2e86a86743b83dbdaeef4c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224694
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>