MB-65725: Revert "MB-65461: Aggregate user activity when fetching user details"
This reverts commit a9c172efef6f0b57a9cfd516a223dcf7847d8184.

The commit was written wrongly assuming that activity_aggregator could
be called locally on all nodes. Since it is only on the orchestrator
node, the call fails on all non-orchestrator nodes.

This breaks the users page in the UI (stuck loading) when connecting
to a non-orchestrator node. Since the fix is non-trivial - the call
would not be as cheap if it went to remote node, so the approach needs
re-evaluating - it is best to revert this change for now, and re-open
the original ticket for further work.

Change-Id: I499c51a9d281b21cc403b4bbe28add7c0b329e99
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224834
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>