MB-64387: num_vbuckets related fixes
The new default values for num_vbuckets should not be used until the
cluster compat mode reaches morpheus. This is now done.

In a mixed version cluster with pre-7.6 nodes, a bucket create
request coming from the pre-7.6 node will not contain a num_vbuckets
property. This wasn't specifiable until 7.6. This coupled with changes
in morpheus to support different default num_vbuckets values (128 for
magma and 1024 for couchstore) led to num_vbuckets not being specified
in the bucket creation request.

This change detects this and fills in num_vbuckets depending on whether
the bucket is magma vs couchstore.

Change-Id: I56c4d309aa88bbfbf6f7b4f958a595a49a01aa58
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224944
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>