MB-65772: Hook for early expiry of OOTB client cert
For testing of rotation and expiry of OOTB client certs, we need a way
to reduce the NotAfter time to a point arbitrarily near in the future.

Adds 'clientCertNotAfterDurationSeconds' to /internalSettings.

After configuring this value, the existing client cert will not be
affected. A new client cert with a new NotAfter can be generated by
re-generating the cluster cert:
POST /controller/regenerateCertificate

Since we automatically regenerate the client cert once expiry is within
7 days, any NotAfterDuration of 604800 or less will immediately trigger
internal client cert rotation (after the cert is regenerated), which
will get repeatedly triggered indefinitely, until the setting is set to
a larger number again.

Change-Id: I8bc087bc66d98365473457fb32bd932570f65a65
Reviewed-on: https://review.couchbase.org/c/ns_server/+/225073
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>