MB-62911: Abolish existing lease after node rename
When a node is renamed from cb.local to its "real" name we abolish the
lease instead of waiting for it to expire. This allows other actions
requiring a lease to begin sooner.

The primary work flow is spinning up a node with a bucket for test case
instances (clusterInit followed by a bucket creation) as quickly as
possible.  The bucket cleanup to assign servers and vbucket map requires
a lease. Abolishing the lease for the rename allows the bucket cleanup
to start sooner.

Change-Id: I943675556bade3578c4803b4a224c672d50e8f3a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/225158
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>