MB-65563: enforce fixedServices defined in config profile
Enforce the set of services on all cluster nodes specified by the
'fixedService' list if defined in the config profile. When configured,
the 'services' parameter for cluster init/server add is no longer
required, but must match the set of fixed services when supplied else
an error will be returned.

Change-Id: I74172f6391901d3e79310bff91864663c0fbbb36
Reviewed-on: https://review.couchbase.org/c/ns_server/+/224941
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Michael Blow <michael.blow@couchbase.com>