MB-65262: Warn when setting soft > hard limit ...
... for cgroups. This configuration is allowed by the kernel but doesn't
actually "mean anything" or provide any useful or unique functionality.
You would just never actually see the soft limit get hit since the hard
limit will always get hit first. So because of this we wanted to still
allow setting this configuration, to keep it 1-to-1 with the kernel, but
wanted to at least provide a warning to the user so they might
understand why it doesn't make sense to do so.

Change-Id: I05b80774f7005c33c2bd207635b1d4b98ccbb595
Reviewed-on: https://review.couchbase.org/c/ns_server/+/225314
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>