MB-65955: Alert for indexes with diverging replicas
When indexes have diverging item counts between replicas, inconsistent
data can be returned by an index scan. This alert is added as an opt-in
option, so that administrators can detect this issue and resolve it.

The detection is handled by the index service and exposed via the
'index_num_diverging_replica_indexes' metric, so ns_server simply has
to check if this metric increases.

Change-Id: I54062b34e506b9f3bc2f3d31b8e33560ce45484e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/225702
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>