MB-60450: Remove mcd connection_limit_mode setting
Memcached ignores the value so allowing a user to change it is
pointless. The parameter will only be removed from the API when the
compat mode moves to morpheus/8.0, allowing the user to continue to set
it, should they wish, til the whole cluster is ugpraded. If the user
sets the value, a morpheus node will ignore it and not pass it to the
memcached config, but other pre-morpheus nodes will. The config will
be removed from the global memcached and the per-node memcached config
overrides on moprheus compat mode upgrade.

Change-Id: Ia4123d57ea369911489f51ac94466bffe183ac36
Reviewed-on: https://review.couchbase.org/c/ns_server/+/225799
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>