MB-65254: Handle websocket handshake response in ns_server
By handling the response in ns_server, we can ensure that the request
gets logged in the http_access.log.

This also ensures that we can avoid responding to the request until we
have added the client to our connection pool, so we can return an error
instead if the pool is full. This is required for MB-65915 which will
be fixed in a subsequent commit.

Change-Id: I1ca29b1e2088c6062c90012ddff341f4b1fad573
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226078
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>