MB-66264: Don't call GET_TELEMETRY on connecting clients
It is possible for the GET_TELEMETRY message to be sent before the
initial HTTP response is sent. This should be prevented by responding
to the request before returning from the gen_server call, ensuring that
the pool won't use the connection until after responding.

Change-Id: Ia096e754c3f29b94aa2ee628e78c460f280aa646
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226295
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>